LayoutTests/svg/animations/animate-currentColor-expected.txt

SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(rect).fill is '#ff0000'
PASS rect.style.fill is ""
PASS getComputedStyle(rect).fill is '#804000'
PASS rect.style.fill is ""
PASS getComputedStyle(rect).fill is '#008000'
PASS rect.style.fill is ""
PASS successfullyParsed is true

TEST COMPLETE
